Provide php-fpm-nginx & rename php
The existing php package becomes php-apache, still providing "php"
though, so anything which reconciles provision aliases will work.

We add a new package "php-fpm-nginx" which declares a dependency upon
nginx and builds php with fpm FastCGI support.

We document in the nginx package that FastCGI is explicitly required.
(We can't constrain this, because of limitations in the nginx build
system).

In all these files, we adjust `${INSTALLPATH}` to become
`${INSTALLPATH:?}` which provides for protection against copy/pasting
the command-lines into a terminal window in which INSTALLPATH is not
set.  That was ... unpleasant.

There are accompanying changes in the continuum-stagers repo, adding
new options to the php stager.

nb: this corresponds to commits 7c8eb41, 24eb9fc & 7b85c13 in the
    continuum repo (not linking, they'll be GC'd), which correspond to
    PR apcera/continuum#1423 -- the location of these files switched
    repo under this PR.