zsh: MD5SUMS file available
Peter Stephenson pointed to <http://www.zsh.org/pub/MD5SUM> which will
be mentioned in future releases (and perhaps made more discoverable) but
it's auto-updated on upload, so doesn't buy a lot of protection (not
even as much as MD5 might provide).  So I've left the more thorough
verification steps in, albeit marked as not necessary.

There is upstream discussion about, minimally, including the release
checksums in a file checked into git as a commit after the release, as
well as switching from MD5.