Fix swapped Python PGP keyids.
Added note pointing to use of 16-char keyids, with rationale.

Python 3.3.5 was released, conflicted with the keyid shown in this
document.  Looks like the 2.7.x and 3.3.x keys were swapped.  I
downloaded:

 * http://www.python.org/ftp/python/2.7.6/Python-2.7.6.tgz.asc
 * http://www.python.org/ftp/python/3.3.3/Python-3.3.3.tgz.asc

and ran `gpg --list-packets` against those signatures, to confirm that
the values matched the set of keyids formerly listed in Verification.md
but that the values were swapped.

Fixed to correct values, given in 0xLong format, and note
reconfirmation.