Zsh 5.0.7 packaged
This is a semi-security fix: there is unexpected evaluation at shell
startup based on values in a couple of specific environment variables,
not all variables.  The _safe_ approach is to treat it as a security
fix.

Verification of source integrity described by me at:
  <http://www.zsh.org/mla/workers/2014/msg01203.html>

Built in Bagel, available as `/apcera/pkg/packages::zsh-5.0.7`