This removes the version field from all build scripts.
The version field within build scripts was never used/leveraged by anything in
the stager. If anything, it was originally meant as metadata for the person
looking at the file. However, this means it confuses people who assume since it
it is there, it is used.

This removes all version fields from build scripts and instead appends the
version to the package name which is used when building the package.